+++ src/Pim/Bundle/WebServiceBundle/Handler/Rest/ProductHandler.php
@@ -44,7 +44,7 @@ class ProductHandler
                 'locales'     => $locales,
                 'channels'    => $channels,
                 'resource'    => $url,
-                'filter_type' => 'pim:external_api:product:view'
+                'filter_type' => 'pim.external_api.product.view'
             ]
         );